Made a db migration for the ScholarshipApplication model. It works fine. Looks bad, but it works.
Basically, its all really ugly and I have some double files. Need to clean it up. I'm not sure how to do that though yet cause I just fucked with a ton of shit till it worked sort of.
Now I need to figure out what worked and clean up the stuff. Then use the clean version to do it on the main branch. But before that I want to adjust and mess with the db structure a bit and the form structure and update some of the form fields.